special case dispatchers

+++ inc/slash/dispatch.h
@@ -10,16 +10,16 @@ int
 sl_responds_to2(sl_vm_t* vm, SLVAL object, SLID idv);
 
 SLVAL
-sl_apply_method(sl_vm_t* vm, SLVAL recv, sl_method_t* method, size_t argc, SLVAL* argv);
+sl_apply_method(sl_vm_t* vm, SLVAL recv, sl_method_t* method, int argc, SLVAL* argv);
 
 SLVAL
-sl_send(sl_vm_t* vm, SLVAL recv, char* id, size_t argc, ...);
+sl_send(sl_vm_t* vm, SLVAL recv, char* id, int argc, ...);
 
 SLVAL
-sl_send_id(sl_vm_t* vm, SLVAL recv, SLID id, size_t argc, ...);
+sl_send_id(sl_vm_t* vm, SLVAL recv, SLID id, int argc, ...);
 
 SLVAL
-sl_send2(sl_vm_t* vm, SLVAL recv, SLID id, size_t argc, SLVAL* argv);
+sl_send2(sl_vm_t* vm, SLVAL recv, SLID id, int argc, SLVAL* argv);
 
 sl_method_t*
 sl_lookup_method(sl_vm_t* vm, SLVAL recv, SLID id);
